Add support for ST2110 payload type (96..127)
Modify SDK, Media Proxy, Mesh Agent, and FFmpeg plugin
to support SMPTE ST2110 payload type, which is typically 96..127.

+++ sdk/include/mesh_conn.h
@@ -78,7 +78,7 @@ public:
             int transport;
 
             std::string pacing;
-            uint32_t payload_type;
+            uint8_t payload_type;
             std::string transportPixelFormat;
         } st2110;
